namespace refactoring and fix test
