+++ hoot-core-test/src/test/cpp/hoot/core/visitors/poi-polygon/PoiPolygonMatchVisitorTest.cpp
@@ -29,7 +29,7 @@
 #include <hoot/core/TestUtils.h>
 #include <hoot/core/conflate/matching/Match.h>
 #include <hoot/core/conflate/matching/MatchThreshold.h>
-#include <hoot/core/conflate/poi-polygon/visitors/PoiPolygonMatchVisitor.h>
+#include <hoot/core/visitors/poi-polygon/PoiPolygonMatchVisitor.h>
 #include <hoot/core/io/OsmXmlReader.h>
 #include <hoot/core/util/Log.h>
 #include <hoot/core/util/MapProjector.h>
